using popt
The popt library exists essentially for parsing command-line
options. It is found superior in many ways when compared to
parsing the argv array by hand or using the getopt functions
getopt() and getopt_long(). Some specific advantages of popt
are: it does not utilize global variables, thus enabling
multiple passes in parsing argv; it can parse an arbitrary
array of argv-style elements, allowing parsing of
command-line-strings from any source; it provides a standard
method of option aliasing; it can exec external option filters;
and, finally, it can automatically generate help and usage
messages for the application.