golang slice
Slices are a key data type in Go, giving a more powerful interface to
sequences than arrays.

Unlike arrays, slices are typed only by the elements they contain (not
the number of elements). To create an empty slice with non-zero length,
use the builtin make.

len(s) returns the length of the slice.

slice supports append(), which returns a slice containing one or more
new values. Note that we need to accept a return value from append as
we may get a new slice value.

slices can also be copy'd.

slice support a "slice" operator with the syntax slice[low:high].