golang errors
In Go it's idiomatic to communicate errors via an explicit, separate return
value. This contrasts with the exception used in languages like Java and Ruby
and the overloaded single result/error value sometimes used in C. Go's approach
makes it easy to see which functions return errors and to handle them using the
same language constructs employed for any other, non-error tasks.

By convention, errors are the last return value and have type error, a built-in
interface.

A nil value in the error position indicates that there was no error.