[cloud-native]: kubernetes backgrounds
During the initial phase of Docker adoption, no out-of-the-box solutions existed
for the orchestration and routing layer, which resulted in users handcrafting
homegrown solutions.

In mid-2014, Google announced the Kubernetes project, an open source system for
automating deployment, scaling, and management of containerized applications.
This provided one of the missing components required in trying to handle running
containers at scale.

Signed-off-by: Zhou Peng <p@ctriple.cn>